Refactored upsample common into .hpp and .cpp and added validations to int factories
